requirements.txt
py-clob-client>=0.9.0
websockets>=12.0
anthropic>=0.40.0
python-telegram-bot>=21.0
pyyaml>=6.0
tenacity>=8.2.0
schedule>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
